pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "enclose"
version = "1.0.1"
description = "A modular document processing pipeline for Markdown to PDF/SVG/PNG conversion with OCR"
authors = ["Tom <info@softreck.dev>"]
readme = "README.md"
license = "MIT"
homepage = "https://github.com/veridock/aggregate"
packages = [
    { include = "enclose" },
]

[tool.poetry.dependencies]
python = "^3.8.1,<4.0"
markdown = "^3.4.4"
weasyprint = "=57.2"
pydyf = "=0.5.0"
reportlab = "^4.0.4"
Pillow = "^10.0.0"
cairosvg = "^2.7.1"
pdf2image = "^1.17.0"
pytesseract = "^0.3.10"
beautifulsoup4 = "^4.12.2"
lxml = "^4.9.3"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-cov = "^4.1.0"
black = "^23.7.0"
isort = "^5.12.0"
flake8 = "^6.1.0"
mypy = "^1.4.1"

[tool.poetry.scripts]
enclose = "enclose.__main__:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
addopts = "-v --cov=enclose --cov-report=term-missing"

[tool.black]
line-length = 88
target-version = ['py38']
include = '\\.pyi?$'

[tool.isort]
profile = "black"
line_length = 88
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
